server_app/requirements.txt
Flask==2.3.3
pyserial==3.5
requests==2.28.1
